Refactor the mui templates
