+++ mui/.prettierignore
@@ -1,2 +1,1 @@
 dist/*
-scripts/mui-icon-builder/templateSvgIcon.js
