--- mui/scripts/mui-icon-builder/modules/waterfall/README.md
@@ -1,3 +1,0 @@
-# Waterfall
-
-A set of utility functions for handling async/await at scale.
