--- mui/scripts/mui-icon-builder/modules/waterfall/sleep.js
@@ -1,7 +1,0 @@
-function sleep(delay = 0) {
-    return new Promise((resolve) => {
-        setTimeout(resolve, delay);
-    });
-}
-
-export default sleep;
