--- mui/scripts/mui-icon-builder/package.json
@@ -1,132 +1,0 @@
-{
-    "name": "mui-icon-builder",
-    "version": "1.0.0-beta",
-    "description": "Library for creating MUI-style icons from custom svgs - this is borrowed from material-ui until it's made available as a package",
-    "main": "./index.js",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/mui/material-ui/tree/master/packages/mui-icons-material"
-    },
-    "author": "MUI",
-    "license": "MIT",
-    "dependencies": {},
-    "scripts": {},
-    "bugs": {
-        "url": "https://github.com/mui/material-ui/issues"
-    },
-    "homepage": "https://github.com/mui/material-ui/tree/master/packages/mui-icons-material",
-    "devDependencies": {
-        "@babel/cli": "7.0.0",
-        "@babel/core": "7.0.0",
-        "@babel/node": "7.0.0",
-        "@babel/plugin-proposal-class-properties": "7.0.0",
-        "@babel/plugin-proposal-object-rest-spread": "7.0.0",
-        "@babel/plugin-transform-object-assign": "7.0.0",
-        "@babel/plugin-transform-runtime": "7.0.0",
-        "@babel/preset-env": "7.0.0",
-        "@babel/preset-react": "7.0.0",
-        "@babel/register": "7.0.0",
-        "@babel/runtime": "7.0.0",
-        "@types/enzyme": "^3.1.4",
-        "@types/react": "^17.0.39",
-        "argos-cli": "^0.1.1",
-        "autoprefixer": "^9.0.0",
-        "autosuggest-highlight": "^3.1.1",
-        "babel-eslint": "^9.0.0",
-        "babel-loader": "^8.0.0",
-        "babel-plugin-istanbul": "^5.0.0",
-        "babel-plugin-module-resolver": "^3.0.0",
-        "babel-plugin-preval": "^2.0.0",
-        "babel-plugin-react-remove-properties": "^0.2.5",
-        "babel-plugin-transform-dev-warning": "^0.1.0",
-        "babel-plugin-transform-react-constant-elements": "^6.23.0",
-        "babel-plugin-transform-react-remove-prop-types": "^0.4.10",
-        "chai": "^4.1.2",
-        "clean-css": "^4.1.11",
-        "clipboard-copy": "^2.0.0",
-        "cross-env": "^5.1.1",
-        "css-loader": "^1.0.0",
-        "doctrine": "^2.0.0",
-        "downshift": "^2.0.0",
-        "dtslint": "^0.3.0",
-        "enzyme": "^3.5.0",
-        "enzyme-adapter-react-16": "^1.3.0",
-        "eslint": "^5.0.0",
-        "eslint-config-airbnb": "^17.0.0",
-        "eslint-import-resolver-webpack": "^0.10.0",
-        "eslint-plugin-babel": "^5.0.0",
-        "eslint-plugin-import": "^2.8.0",
-        "eslint-plugin-jsx-a11y": "^6.0.3",
-        "eslint-plugin-mocha": "^5.0.0",
-        "eslint-plugin-react": "^7.4.0",
-        "eslint-plugin-spellcheck": "^0.0.10",
-        "fg-loadcss": "^2.0.1",
-        "file-loader": "^2.0.0",
-        "fs-extra": "^7.0.0",
-        "glob": "^7.1.2",
-        "gm": "^1.23.0",
-        "isomorphic-fetch": "^2.2.1",
-        "jsdom": "^16.5.0",
-        "jss-rtl": "^0.2.1",
-        "karma": "^4.2.0",
-        "karma-browserstack-launcher": "^1.3.0",
-        "karma-chrome-launcher": "^2.2.0",
-        "karma-mocha": "^1.3.0",
-        "karma-mocha-reporter": "^2.2.5",
-        "karma-sourcemap-loader": "^0.3.7",
-        "karma-webpack": "^3.0.0",
-        "lz-string": "^1.4.4",
-        "mkdirp": "^0.5.0",
-        "mocha": "^5.0.0",
-        "mustache": "^3.0.0",
-        "next": "^12.1.0",
-        "nyc": "^13.0.0",
-        "postcss": "^7.0.0",
-        "prettier": "^1.8.2",
-        "puppeteer": "^1.5.0",
-        "raw-loader": "^0.5.1",
-        "react": "^17.0.2",
-        "react-autosuggest": "^9.3.2",
-        "react-docgen": "^3.0.0-beta10",
-        "react-dom": "^17.0.2",
-        "react-inspector": "^2.2.2",
-        "react-jss": "^8.1.0",
-        "react-markdown": "^3.4.1",
-        "react-number-format": "^3.0.2",
-        "react-redux": "^5.0.6",
-        "react-router": "^4.2.0",
-        "react-router-dom": "^4.2.2",
-        "react-select": "^2.0.0",
-        "react-swipeable-views": "^0.13.0",
-        "react-test-renderer": "^16.1.1",
-        "react-text-mask": "^5.0.2",
-        "recast": "^0.15.0",
-        "recharts": "^1.1.0",
-        "recompose": "0.28.0 - 0.30.0",
-        "redux": "^4.0.0",
-        "redux-logger": "^3.0.6",
-        "rimraf": "^2.6.2",
-        "rollup": "^0.66.0",
-        "rollup-plugin-babel": "^4.0.0-beta.8",
-        "rollup-plugin-commonjs": "^9.1.3",
-        "rollup-plugin-node-globals": "^1.2.1",
-        "rollup-plugin-node-resolve": "^3.3.0",
-        "rollup-plugin-size-snapshot": "^0.7.0",
-        "rollup-plugin-uglify": "^6.0.0",
-        "sinon": "^6.0.0",
-        "size-limit": "~0.20.0",
-        "svgo": "^1.0.5",
-        "temp": "^0.8.3",
-        "tslint": "^5.11.0",
-        "typescript": "^3.0.0",
-        "url-loader": "^1.0.1",
-        "vrtest": "^0.2.0",
-        "webfontloader": "^1.6.28",
-        "webpack": "4.20.2",
-        "webpack-bundle-analyzer": "^3.0.0",
-        "webpack-cli": "^3.1.0",
-        "workbox-build": "^3.0.1",
-        "yargs": "^12.0.0"
-    },
-    "keywords": []
-}
