+++ mui/svgr-template.js
@@ -1,9 +1,11 @@
 const template = ({ imports, interfaces, componentName, jsx, exports }, { tpl }) => {
-    return tpl`${imports}; 
+    return tpl`
+    ${imports};
     import { createSvgIcon } from '@mui/material/utils';
     ${interfaces};
-    const ${componentName} = createSvgIcon(${jsx.children}, '${componentName}');
-    ${exports};`;
+    const ${componentName} = createSvgIcon(React.createElement(React.Fragment, {}, ${jsx.children}), '${componentName}');
+    ${exports};
+    `;
 };
 
 module.exports = template;
